Remove overlay and revert X to hamburger icon when popup disappears (#3958)
the `.hidden-links` popup disappears after you move the cursor out of it.
However, the white overlay on the screen does not disappear, as well as the `x` nav button on the top-right, remains as is.
Now when you click anywhere on the screen the popup appears again, and the `x` nav button turns back to a hamburger menu button.

And when you click on the hamburger menu, the popup disappears. now it's happening the exact opposite.

To remedy this, the pop up menu as well as the `x` nav button should also return back to normal once the popup disappears.

This commit fixes this issue.